CentralizedTelescope: Moving algo.rs in proof.rs
